requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.2
python-multipart>=0.0.9
pydantic>=2.9.0
regex>=2024.9.11
